build correctly during development
This runs the babel build script when installing the package as a dependency. Prior to this, installing and configuring this package as a dependency results in a silent failure (or success, depending on how you look at it) when running `gatsby develop`. For some reason, gatsby just ignores this at the moment, throws no error and otherwise compiles successfully.